Refactor src director for redux
