+++ src/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import App from './App';
-import Books from './pages/Books';
-import Categories from './pages/Categories';
+import Books from './redux/books/Books';
+import Categories from './redux/categories/Categories';
 import Header from './components/Header';
 import './css/index.css';
